Merge

+++ shelfy/app/src/main/java/com/example/shelfy/data/remote/responses/Books.kt
@@ -3,5 +3,5 @@ package com.example.shelfy.data.remote.responses
 data class Books(
     val items: List<Item> = emptyList(),
     val kind: String = "",
-    val totalItems: Int = 0
+    val totalItems: Long = 0
 )
