+++ shelfy/app/src/main/java/com/example/shelfy/data/remote/responses/Offer.kt
@@ -1,7 +1,7 @@
 package com.example.shelfy.data.remote.responses
 
 data class Offer(
-    val finskyOfferType: Int = 0,
+    val finskyOfferType: Long = 0,
     val listPrice: ListPriceX = ListPriceX(),
     val retailPrice: RetailPrice = RetailPrice()
 )
